[OT-329] [REFACTOR]: ViewProgressBar 컴포넌트 shared로 이동
