+++ src/entities/home/components/index.ts
@@ -1,10 +1,9 @@
-export { default as MainCarousel } from './MainCarousel';
-export { default as ContentCarousel } from './ContentCarousel';
-export { default as TrendingCarousel } from './TrendingCarousel';
+export { default as MainCarousel } from "./MainCarousel";
+export { default as ContentCarousel } from "./ContentCarousel";
+export { default as TrendingCarousel } from "./TrendingCarousel";
 export { default as RecommendCarousel } from "./RecommendCarousel";
 export { default as RecommendTagsCarousel } from "./RecommendTagsCarousel";
 export { default as RecommendTagsSection } from "./RecommendTagsSection";
 export { default as HistoryCarousel } from "./HistoryCarousel";
 export { default as CustomRecommendCarousel } from "./CustomRecommendCarousel";
-export { default as ViewProgressBar } from "./ViewProgressBar";
-export { default as AiCardSlide } from "./AiCardSlide"
+export { default as AiCardSlide } from "./AiCardSlide";
