+++ src/shared/components/ViewProgressBar.tsx
