+++ src/shared/components/index.ts
@@ -1,12 +1,13 @@
-export { BackButton } from './BackButton';
-export { Badge } from './Badge';
-export { CommonButton } from './CommonButton';
-export { default as ConfirmModal } from './ConfirmModal';
-export { Input } from './Input';
-export { InteractionButton } from './InteractionButton';
-export { ScrollEdgeButton } from './ScrollEdgeButton';
-export { Toggle } from './Toggle';
+export { default as ViewProgressBar } from "./ViewProgressBar";
+export { BackButton } from "./BackButton";
+export { Badge } from "./Badge";
+export { CommonButton } from "./CommonButton";
+export { default as ConfirmModal } from "./ConfirmModal";
+export { Input } from "./Input";
+export { InteractionButton } from "./InteractionButton";
+export { ScrollEdgeButton } from "./ScrollEdgeButton";
+export { Toggle } from "./Toggle";
 
-export type { CommonButtonProps } from './CommonButton';
-export type { ToggleProps } from './Toggle';
-export type { ScrollEdgeButtonProps } from './ScrollEdgeButton';
+export type { CommonButtonProps } from "./CommonButton";
+export type { ToggleProps } from "./Toggle";
+export type { ScrollEdgeButtonProps } from "./ScrollEdgeButton";
